backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg  # for Postgres
python-dotenv
langchain
langchain-groq
langgraph
langchain-core